agent/requirements.txt
livekit-agents>=0.8.12
livekit-plugins-silero
gradio-client